% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNARclust.R
\docType{data}
\name{echo}
\alias{echo}
\title{Echocardiogram data set}
\format{
A data frame with 132 observations on 13 variables
(more details on this data set are presented in http://archive.ics.uci.edu/ml/datasets/Echocardiogram).
}
\description{
All the patients suffered heart attacks at some point in the past. Some are still alive and some are not. The survival and still-alive variables, when taken together, indicate whether a patient survived for at least one year following the heart attack.
}
\details{
This data set arise from the UCI machine learning repository (more details on this data set are presented http://archive.ics.uci.edu/ml/datasets/Echocardiogram)
}
\examples{
data(echo)
}
\references{
Salzberg, S. (1988). Exemplar-based learning: Theory and implementation (Technical Report TR-10-88). Harvard University, Center for Research in Computing Technology, Aiken Computation Laboratory (33 Oxford Street; Cambridge, MA 02138).
}
\keyword{datasets}
