% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNARclust.R
\docType{package}
\name{MNARclust-package}
\alias{MNARclust-package}
\alias{MNARclust}
\title{MNARclust.}
\description{
Clustering method to analyze continuous or mixed-type data with missingness. The missingness mechanism can be non ignorable. The approach considers a semi-parametric mixture model.
}
\details{
\tabular{ll}{
  Package: \tab MNARclust\cr
  Type: \tab Package\cr
  Version: \tab 1.1.0\cr
  Date: \tab 2021-12-01\cr
  License: \tab GPL-3\cr
  LazyLoad: \tab yes\cr
}
}
\keyword{package}
